pyproject.toml
[tool.poetry]
name = "quotient-bot"
version = "0.1.0"
description = ""
authors = ["deadaf <gahlyanrohit18@gmail.com>"]
license = "MPL 2.0"
readme = "README.md"
packages = [{include = "quotient_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.0"
tortoise-orm = "^0.19.3"
colorama = "^0.4.6"
asyncpg = "^0.27.0"
dateparser = "^1.1.8"
aerich = "^0.7.1"
parsedatetime = "^2.6"
pillow = "^9.5.0"
psutil = "^5.9.5"
pynacl = "^1.5.0"
jishaku = "^2.5.1"
humanize = "^4.6.0"
pygit2 = "^1.12.1"
aiocache = {extras = ["msgpack"], version = "^0.12.1"}
ujson = "^5.8.0"
imagehash = "^4.3.1"
python-socketio = "^5.8.0"
aiohttp-asgi = "^0.5.2"
fastapi = "^0.98.0"
python-multipart = "^0.0.6"
lru-dict = "^1.2.0"
jinja2 = "^3.1.2"
prettytable = "^3.8.0"
uvicorn = "^0.35.0"
replit = "^4.1.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[tool.black]
line-length = 122

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

